+++ tests/testthat/test-download_data_macro_predictors.R
@@ -3,7 +3,7 @@ test_that("download_data_macro_predictors gracefully handles broken url", {
   skip_on_cran()
   expect_message(
     download_data_macro_predictors(
-      "macro_predictors_monthly",
+      "monthly",
       start_date = "2020-01-01",
       end_date = "2022-12-01",
       sheet_id = "test"
@@ -16,7 +16,7 @@ test_that("download_data_macro_predictors gracefully handles broken url", {
 test_that("download_data_macro_predictors handles start_date after end_date", {
   expect_error(
     download_data_macro_predictors(
-      "macro_predictors_monthly",
+      dataset = "monthly",
       start_date = "2021-12-31",
       end_date = "2020-01-01"
     ),