+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 * Added `download_data_hf()` and support for `type = "hf_high_frequency_sp500"`.
+* Deprecated `type` parameter in favor of `domain` and `dataset`.
 
 ## Improvements